+++ src/LaravelDDDServiceProvider.php
@@ -2,15 +2,6 @@
 
 namespace Lunarstorm\LaravelDDD;
 
-use Lunarstorm\LaravelDDD\Commands\DomainActionMakeCommand;
-use Lunarstorm\LaravelDDD\Commands\DomainBaseModelMakeCommand;
-use Lunarstorm\LaravelDDD\Commands\DomainBaseViewModelMakeCommand;
-use Lunarstorm\LaravelDDD\Commands\DomainDtoMakeCommand;
-use Lunarstorm\LaravelDDD\Commands\DomainFactoryMakeCommand;
-use Lunarstorm\LaravelDDD\Commands\DomainModelMakeCommand;
-use Lunarstorm\LaravelDDD\Commands\DomainValueObjectMakeCommand;
-use Lunarstorm\LaravelDDD\Commands\DomainViewModelMakeCommand;
-use Lunarstorm\LaravelDDD\Commands\InstallCommand;
 use Spatie\LaravelPackageTools\Package;
 use Spatie\LaravelPackageTools\PackageServiceProvider;
 
@@ -27,30 +18,31 @@ class LaravelDDDServiceProvider extends PackageServiceProvider
             ->name('laravel-ddd')
             ->hasConfigFile()
             ->hasCommands([
-                InstallCommand::class,
-                DomainModelMakeCommand::class,
-                DomainFactoryMakeCommand::class,
-                DomainBaseModelMakeCommand::class,
-                DomainDtoMakeCommand::class,
-                DomainValueObjectMakeCommand::class,
-                DomainViewModelMakeCommand::class,
-                DomainBaseViewModelMakeCommand::class,
-                DomainActionMakeCommand::class,
-                \Lunarstorm\LaravelDDD\Commands\DomainCastMakeCommand::class,
-                \Lunarstorm\LaravelDDD\Commands\DomainChannelMakeCommand::class,
-                \Lunarstorm\LaravelDDD\Commands\DomainConsoleMakeCommand::class,
-                \Lunarstorm\LaravelDDD\Commands\DomainEventMakeCommand::class,
-                \Lunarstorm\LaravelDDD\Commands\DomainExceptionMakeCommand::class,
-                \Lunarstorm\LaravelDDD\Commands\DomainJobMakeCommand::class,
-                \Lunarstorm\LaravelDDD\Commands\DomainListenerMakeCommand::class,
-                \Lunarstorm\LaravelDDD\Commands\DomainMailMakeCommand::class,
-                \Lunarstorm\LaravelDDD\Commands\DomainNotificationMakeCommand::class,
-                \Lunarstorm\LaravelDDD\Commands\DomainObserverMakeCommand::class,
-                \Lunarstorm\LaravelDDD\Commands\DomainPolicyMakeCommand::class,
-                \Lunarstorm\LaravelDDD\Commands\DomainProviderMakeCommand::class,
-                \Lunarstorm\LaravelDDD\Commands\DomainResourceMakeCommand::class,
-                \Lunarstorm\LaravelDDD\Commands\DomainRuleMakeCommand::class,
-                \Lunarstorm\LaravelDDD\Commands\DomainScopeMakeCommand::class,
+                Commands\InstallCommand::class,
+                Commands\DomainListCommand::class,
+                Commands\DomainModelMakeCommand::class,
+                Commands\DomainFactoryMakeCommand::class,
+                Commands\DomainBaseModelMakeCommand::class,
+                Commands\DomainDtoMakeCommand::class,
+                Commands\DomainValueObjectMakeCommand::class,
+                Commands\DomainViewModelMakeCommand::class,
+                Commands\DomainBaseViewModelMakeCommand::class,
+                Commands\DomainActionMakeCommand::class,
+                Commands\DomainCastMakeCommand::class,
+                Commands\DomainChannelMakeCommand::class,
+                Commands\DomainConsoleMakeCommand::class,
+                Commands\DomainEventMakeCommand::class,
+                Commands\DomainExceptionMakeCommand::class,
+                Commands\DomainJobMakeCommand::class,
+                Commands\DomainListenerMakeCommand::class,
+                Commands\DomainMailMakeCommand::class,
+                Commands\DomainNotificationMakeCommand::class,
+                Commands\DomainObserverMakeCommand::class,
+                Commands\DomainPolicyMakeCommand::class,
+                Commands\DomainProviderMakeCommand::class,
+                Commands\DomainResourceMakeCommand::class,
+                Commands\DomainRuleMakeCommand::class,
+                Commands\DomainScopeMakeCommand::class,
             ]);
 
         // Enum generator only in Laravel 11